requirements.txt
annotated-types==0.6.0
anthropic==0.20.0
anyio==4.3.0
asgiref==3.7.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.11
django-crispy-forms==2.1
django-environ==0.11.2
django-tastypie==0.14.6
djangorestframework==3.14.0
environ==1.0
exceptiongroup==1.2.0
fcm-django==2.0.1
filelock==3.13.1
firebase-admin==6.5.0
fsspec==2024.3.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
msgpack==1.0.8
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-mimeparse==1.6.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
